made some progress on synthSong, now have a sounding sine whoo

+++ CSD2b/synthSong/codebase/sine.cpp
@@ -1,7 +1,7 @@
 #include "sine.h"
 #include "math.h"
 
-Sine::Sine (double frequency, double samplerate)
+Sine::Sine (float frequency, float samplerate)
   : Oscillator (frequency, samplerate) {
 
   #if DEBUG
